requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.19.2
stripe==4.1.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.0.0
